backed/requirements.txt
fastapi>=0.110
python-multipart>=0.0.9
PyPDF2>=3.0
python-docx>=1.1
uvicorn>=0.27
httpx>=0.26
